requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
yfinance>=0.2.0
scipy>=1.10.0
python-dateutil>=2.8.0
pyyaml>=6.0
requests>=2.31.0
openpyxl>=3.1.0
